Only check the generated expression when checking for duplicate expressions with ambiguous symbols
Before it was checking the LexToken that represented it as well. If the generated expression is the same we expect the length of the stream to be the same.
If this isn't ever the case then we can check the length cheaply - as they both start from the same slice the contents of the stream are not important anyway.